Wprowadzono mechanizm dodawania pliku, zmodyfikowanobazę danych oraz typy, dopasowując je do rzeczywistych wymagań projektowych. Działające dodawanie plików i artykułów

+++ Horyzonty_Nauki.Application/Attachments/AttachmentDto.cs
@@ -16,7 +16,7 @@ namespace Horyzonty_Nauki.Application.Attachments
         public string File_name { get; set; }
         public string File_type { get; set; }
 
-        public int File_size { get; set; }
+        public long File_size { get; set; }
 
         public string File_path { get; set; }
     }
